Indicator 및 HighlightedDiary 컴포넌트화
