+++ app/src/main/java/com/sopt/smeem/presentation/compose/components/SmeemPagerIndicator.kt
@@ -1,0 +1,40 @@
+package com.sopt.smeem.presentation.compose.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import com.sopt.smeem.presentation.compose.theme.black
+import com.sopt.smeem.presentation.compose.theme.gray300
+
+@Composable
+fun SmeemPagerIndicator(
+    currentPage: Int,
+    pageCount: Int,
+    modifier: Modifier = Modifier,
+    indicatorSize: Dp = 8.dp,
+    indicatorSpacing: Dp = 8.dp
+) {
+    Row(
+        modifier = modifier,
+        horizontalArrangement = Arrangement.spacedBy(indicatorSpacing),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        for (i in 0 until pageCount) {
+            Box(
+                modifier = Modifier
+                    .size(indicatorSize)
+                    .background(
+                        color = if (i == currentPage) black else gray300,
+                        shape = androidx.compose.foundation.shape.CircleShape
+                    )
+            )
+        }
+    }
+}
